Update getWeatherImpact() return type as int

+++ src/com/aslihanhasar/JavaProjects/homeworks/fourthWeek/pokemonGameSimulation/model/WeatherType.java
@@ -17,7 +17,7 @@ public enum WeatherType {
         this.randomValue=randomValue;
     }
 
-    public double getWeatherImpact() {
+    public int getWeatherImpact() {
         return weatherImpact;
     }
